Crate orbiting_sand

Modules

entities

This module contains all the top level bundles. These are entities that are used in the game.

gui

This module contains all the GUI related code. Things that are drawn to via screen coordinates rather than world coordinates.

physics

This module contains all the physics related code.

Functions

main 🔒

planet_only_setup 🔒

Creates just a planet

solar_system_setup 🔒

Creates a solar system with a sun, earth, and a bunch of asteroids.