Module orbiting_sand::entities::celestials::celestial

Celestial entities and their data A celestial is a large body in space, such as a planet or star

Structs

CelestialBuilder

Create a celestial using a builder pattern

CelestialChunkIdk

A component that represents a chunk by its index in the directory

CelestialData

Acts as a cache for a polar mesh’s meshes and textures

CelestialDataPlugin

A plugin that adds the CelestialData system

CelestialOutline

Identifies the mesh which draws the celestials chunk outlines

CelestialWireframe

Identifies the mesh which draws the celestial cell wireframes

FallingSandMaterial

Put this alongside the mesh that represents the falling sand itself