Module orbiting_sand::gui::camera

The bevy camera for the game

Structs

BackgroundLayer1

A layer behind the game. Z-index = -1

CameraPlugin

The plugin for the camera system

CelestialIdx

A component that allows us to enumerate over all the celestials

MainCamera

Used to help identify our main camera

OverlayLayer1

A layer in front of the game. Z-index = 1

OverlayLayer2

A layer in front of the game. Z-index = 2

OverlayLayer3

A layer in front of the game. Z-index = 3

SelectCelestial

An event that indicates that a celestial has been selected by the user

Functions

focus_celestial 🔒

Parent the camera to the celestial

rect_add 🔒

Add a vector to every corner of a rectangle

rect_overlaps 🔒

Check if two rectangles overlap