Module orbiting_sand::physics::fallingsand

Falling sand physics engine. Relates to creating planets and other celestial bodies. And managing their cellular automata.

Modules

convolution

The convolution module contains the logic for getting elements from locations relative to another element in a way that is convienient for parallel processing of chunks.

data

The data module contains the data structures for the falling sand physics engine. These are where the actual grids of elements are stored.

elements

This module contains all the elements that are used in the game.

mesh

The mesh module contains the wireframe layout of a Celestial.

util