Module orbiting_sand::physics::fallingsand::convolution::behaviors

Element Behavior API

This module contains the behaviors of the convolution It exports the ElementGridConvolutionNeighbors struct which contains both the indexes and the grids of the neighbors It contains getters for the chunks based on index or identifier and both getters and setters by element It also allows the user to get the element “below” or “left”/“right” of a given element This is its main purpose, in defining element behaviors, being able to check and change the state of the elements around it. This is the main way that the elements interact with each other. Within the same chunk this is very easy, but between chunks the behavior is very complex:

What does “down” mean if the resolution of cells below you is half the resolution of cells in your chunk?

What does “up” mean if the resolution of cells above you is double the resolution of cells in your chunk?

What if you want to go diagonally between chunks

What are the boundaries of the chunks? Have you accessed a value you don’t have ownership of?

All of that is handled automagically by this API.

Structs

ElementGridConvolutionNeighbors

This is the main struct exported by this module It contains all the neighbors of a chunk in the convolution It does not contain the target chunk itself, as this usually violates several borrow checker rules

ElementGridConvolutionNeighborsIntoIter

Iteration We are going to implement into interation on the Neighbors so that unpackaging is easier To do this we will use the into_hashmap method on the neighbor grids and the iter method on the neighbor indexes taking from the hashmap on each iteration of the iter

Enums

GetChunkErr

Errors for the getter methods