Module orbiting_sand::physics::fallingsand::convolution::neighbor_grids

Actual data storage types for the convolution

Structs

ConvOutOfBoundsError

Defines when the user has simply exceeded the bounds of the convolution

ElementGridConvolutionNeighborGrids

The main type exported by this module Contains all the neighbor grids for the convolution Check out the super::neighbor_identifiers::ConvolutionIdentifier and super::neighbor_indexes::ElementGridConvolutionNeighborIdxs documentation for more information

Enums

BottomNeighborGrids

Bottom neighbor grids in the convolution Check out the super::neighbor_identifiers::BottomNeighborIdentifier and super::neighbor_indexes::BottomNeighborIdxs documentation for more information

LeftRightNeighborGrids

Left and Right neighbor grids in the convolution Check out the super::neighbor_identifiers::LeftRightNeighborIdentifier and super::neighbor_indexes::LeftRightNeighborIdxs documentation for more information documentation for more information

TopNeighborGrids

Top neighbor grids in the convolution Check out the super::neighbor_identifiers::TopNeighborIdentifier and super::neighbor_indexes::TopNeighborIdxs documentation for more information