Module orbiting_sand::physics::fallingsand::convolution::neighbor_identifiers

Identifiers for the different locations chunks can be in the convolution

Structs

BottomNeighborIdentifierChunkDoublingIter

An iterator over the variants of BottomNeighborIdentifierChunkDoubling

BottomNeighborIdentifierIter

An iterator over the variants of BottomNeighborIdentifier

BottomNeighborIdentifierNormalIter

An iterator over the variants of BottomNeighborIdentifierNormal

ConvolutionIdx

Identifies a coordinate on an element grid, and then uniquely identifies the chunk in the convolution that it is in

TopNeighborIdentifierChunkDoublingIter

An iterator over the variants of TopNeighborIdentifierChunkDoubling

TopNeighborIdentifierIter

An iterator over the variants of TopNeighborIdentifier

TopNeighborIdentifierNormalIter

An iterator over the variants of TopNeighborIdentifierNormal

Enums

BottomNeighborIdentifier

Identifies a chunk is one of the bottom neighbors in the convolution Check out the super::neighbor_grids::BottomNeighborGrids and super::neighbor_indexes::BottomNeighborIdxs documentation for more information

BottomNeighborIdentifierChunkDoubling

Identifies a chunk doubling layer transition (or halfing in this case because you are going down) One of these will be directly below you, and be bigger than you off to one direction Whereas the other will be diagonally below you This depends on if your crate::physics::fallingsand::util::vectors::ChunkIjkVector has a k value which is even or odd If it is even, then the bl will be directly below you, and you will be straddling its right side If it is odd, then the br will be directly below you, and you will be straddling its left side

BottomNeighborIdentifierNormal

Identifies that the bottom neighbor is not part of a chunk doubling layer transition However, remember that it still may have half the number of cells tangentially

ConvolutionIdentifier

The main type exported by this module An enum that identifies the location in the structure of the convolution Check out the super::neighbor_grids::ElementGridConvolutionNeighborGrids and super::neighbor_indexes::ElementGridConvolutionNeighborIdxs documentation for more information

LeftRightNeighborIdentifier

Identifies a chunk is the left or right neighbor in the convolution Check out the super::neighbor_grids::LeftRightNeighborGrids and super::neighbor_indexes::LeftRightNeighborIdxs documentation for more information

TopNeighborIdentifier

Identifies a chunk is one of the top neighbors in the convolution Check out the super::neighbor_grids::TopNeighborGrids and super::neighbor_indexes::TopNeighborIdxs documentation for more information

TopNeighborIdentifierChunkDoubling

Identifies a chunk doubling layer transition

TopNeighborIdentifierNormal

Identifies that the bottom neighbor is not part of a chunk doubling layer transition However, remember that it still may have half the number of cells tangentially