Module orbiting_sand::physics::fallingsand::data::element_directory

Structs

ElementGridDir

An element grid directory is like a coordinate directory, but for element grids It follow the same layer structure There is a coordinate directory at the root, but also each ElementGrid has its own copy of the chunk coordinates associated with it for convenience

Parallel 🔒

Useful for indicating at compile time that an iterable should be ran in parallel

ProcessTargets 🔒

Each “item” in this struct gets generated by the cooresponding calculate_ith_*_targets function These each need to be spread out into 9 sets, one for each frame Each of these is ran in sequence relative to each other Whether they are ran in sequence based on their hashsets is different for each name and usually described in the corresponding functions documentation

Sequential 🔒

Useful for indicating at compile time that an iterable should be ran in sequence

Textures

A struct of textures for use in rendering These are options so you can take them out of the struct and use them elsewhere

Constants

FRAMES_PER_FULL_PROCESS 🔒

The number of frames it takes to fully process the directory

Functions

calculate_ith_has_different_k_bottom_neighbor_targets 🔒

This calculates the targets for the edge case where the bottom layer has a different number of tangential chunkss than this layer. In this case we still need to maintain a j step of 3, but we need to process all k’s sequentially.

calculate_ith_has_single_bottom_neighbor_targets 🔒

This really only calculates the core now TODO: Maybe consider removing this

calculate_ith_standard_convolution_targets 🔒

This calculates the convolution targets you would get by standard iteration over 3x3 convolution kernels. It excludes known edge cases, like the bottom of a layer where there is a reduction in tangential chunkss. Run this over 0..9 frame_nb to get the targets for each frame

pregen_process_targets 🔒

Pre calculate all the chunk idx’s we need to process each frame. We pregenerate these so that we can test them and so that we don’t waste time recalculating them