Struct orbiting_sand::physics::fallingsand::data::element_directory::ElementGridDir

pub struct ElementGridDir {
    coords: CoordinateDir,
    chunks: Vec<Grid<Option<ElementGrid>>>,
    process_targets: ProcessTargets,
    process_count: usize,
    total_mass: Mass,
}

An element grid directory is like a coordinate directory, but for element grids It follow the same layer structure There is a coordinate directory at the root, but also each ElementGrid has its own copy of the chunk coordinates associated with it for convenience

Fields
coords: CoordinateDirchunks: Vec<Grid<Option<ElementGrid>>>process_targets: ProcessTargetsprocess_count: usizetotal_mass: Mass
Implementations

impl ElementGridDir

pub fn new_empty(coords: CoordinateDir) -> Self

pub fn new_checkerboard( coords: CoordinateDir, fill0: &dyn Element, fill1: &dyn Element ) -> Self

fn get_chunk_top_neighbors(&self, coord: ChunkIjkVector) -> TopNeighborIdxs

fn get_chunk_left_right_neighbors( &self, coord: ChunkIjkVector ) -> LeftRightNeighborIdxs

fn get_chunk_bottom_neighbors( &self, coord: ChunkIjkVector ) -> BottomNeighborIdxs

fn get_chunk_neighbors( &self, coord: ChunkIjkVector ) -> ElementGridConvolutionNeighborIdxs

pub fn package_coordinate_neighbors( &mut self, coord: ChunkIjkVector ) -> Result<ElementGridConvolutionNeighbors, String>

fn package_convolutions( &mut self, target_chunk_coords: HashSet<ChunkIjkVector> ) -> Result<(Vec<ElementGridConvolutionNeighbors>, Vec<ElementGrid>), Vec<ChunkIjkVector>>

fn unpackage_convolution( &mut self, target: ElementGrid, conv: ElementGridConvolutionNeighbors )

The reverse of the package_convolutions function Puts the chunks back into the directory exactly where they were taken from This is easy because all elementgrids contain a coordinate

fn unpackage_convolutions( &mut self, convolutions: Vec<ElementGridConvolutionNeighbors>, target_chunks: Vec<ElementGrid> )

Unpackages more than one convolution at a time

fn get_unprocessed_chunk_idxs(&self) -> Vec<ChunkIjkVector>

Using the already_processed flag, get all the chunks that have not been processed yet

fn unlock_all_chunks(&mut self)

Sets the already_processed flag to false for all chunks

pub fn process(&mut self, current_time: Clock)

Do one iteration of processing on the grid There are four passes in total, each call does one pass The passes ensure that no two adjacent elementgrids are processed at the same time This is important because elementgrids can effect one another at a maximum range of the size of one elementgrid.

pub fn recalculate_everything(&mut self)

Recalculates all the saved values

pub fn process_full(&mut self, current_time: Clock)

Run process FRAMES_PER_FULL_PROCESS times

pub fn process_single_chunk( &mut self, current_time: Clock, coord: ChunkIjkVector )

Process a single chunk and its neighbors, mostly used for unit testing Also single threaded so should be good for debugging and tracing

pub fn get_updated_target_textures(&self) -> HashMap<ChunkIjkVector, Textures>

Gets the textures of the targets updated in the last call to process

fn process_sequence( &mut self, targets: Sequential<HashSet<ChunkIjkVector>>, current_time: Clock )

fn process_parallel( &mut self, targets: Parallel<HashSet<ChunkIjkVector>>, current_time: Clock )

pub fn get_num_chunks(&self) -> usize

Get the number of chunks from the coordinate directory

pub fn get_total_num_cells(&self) -> usize

pub fn get_total_mass(&self) -> Mass

Get the total mass of the directory

fn recalculate_total_mass(&mut self)

Recalculate the total mass of the directory

pub fn calc_total_mass(chunks: &mut Vec<Grid<Option<ElementGrid>>>) -> Mass

Calculate the maximum temperature in the directory

pub fn get_chunk_by_chunk_ijk(&self, coord: ChunkIjkVector) -> &ElementGrid

Get the maximum temperature in the directory Gets the chunk at the given index Errors if it is currently borrowed

pub fn get_chunk_by_chunk_ijk_mut( &mut self, coord: ChunkIjkVector ) -> &mut ElementGrid

Gets the chunk at the given index mutably Errors if it is currently borrowed

pub fn get_element(&self, coord: IjkVector) -> &Box<dyn Element>

pub fn set_element( &mut self, coord: IjkVector, element: Box<dyn Element>, current_time: Clock )

pub fn get_coordinate_dir(&self) -> &CoordinateDir

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn get_textures(&self) -> HashMap<ChunkIjkVector, Textures>

Get all textures

fn get_textures_filtered( &self, filter: &[Grid<bool>] ) -> HashMap<ChunkIjkVector, Textures>

Where filter is true, get the textures