Struct orbiting_sand::physics::fallingsand::data::element_grid::ElementGrid

pub struct ElementGrid {
    grid: Grid<Box<dyn Element>>,
    coords: ChunkCoords,
    total_mass: Mass,
    already_processed: bool,
    last_set: Clock,
}

An element grid is a 2D grid of elements tied to a chunk

Fields
grid: Grid<Box<dyn Element>>coords: ChunkCoordstotal_mass: Mass

Some low resolution data about the world

already_processed: bool

The minimum temperature in the chunk that is not 0 The maximum temperature in the chunk This deals with a lock during convolution

last_set: Clock

This deals with whether or not the element grid needs to be processed or if it hasn’t seen any changes since the last frame maybe you can skip it

Implementations

impl ElementGrid

pub fn new_empty(chunk_coords: ChunkCoords) -> Self

Creates a new element grid with the given chunk coords and fills it with vacuum

pub fn new_filled(chunk_coords: ChunkCoords, fill: &dyn Element) -> Self

Creates a new element grid with the given chunk coords and fills it with the given element

impl ElementGrid

pub fn get_already_processed(&self) -> bool

pub fn set_already_processed(&mut self, already_processed: bool)

pub fn set_already_processed_deduplicated( &mut self, already_processed: bool ) -> Result<(), String>

Sets the already processed flag and errors if it is set to the same value twice

pub fn get_last_set(&self) -> Clock

pub fn get_chunk_coords(&self) -> &ChunkCoords

pub fn get_grid(&self) -> &Grid<Box<dyn Element>>

pub fn get_total_mass(&self) -> Mass

Does not calculate the total mass, just gets the set value of it

pub fn get_process_unneeded(&self, current_time: Clock) -> bool

Get the maximum temperature in the directory Calculate the maximum temperature in the directory Does not calculate the total mass, just gets the set value of it

impl ElementGrid

Public modifiers for the element grid

pub fn get(&self, jk: JkVector) -> &Box<dyn Element>

pub fn checked_get( &self, jk: JkVector ) -> Result<&Box<dyn Element>, GridOutOfBoundsError>

pub fn get_mut(&mut self, jk: JkVector) -> &mut Box<dyn Element>

pub fn set(&mut self, jk: JkVector, element: Box<dyn Element>, time: Clock)

pub fn replace( &mut self, jk: JkVector, element: Box<dyn Element>, time: Clock ) -> Box<dyn Element>

impl ElementGrid

Proceedural generation helpers

pub fn fill(&mut self, element: ElementType)

Fill the grid with the given element

impl ElementGrid

Handle processing

pub fn process( &mut self, coord_dir: &CoordinateDir, element_grid_conv_neigh: &mut ElementGridConvolutionNeighbors, current_time: Clock )

Do one iteration of processing on the grid

fn process_elements( &mut self, coord_dir: &CoordinateDir, element_grid_conv_neigh: &mut ElementGridConvolutionNeighbors, current_time: Clock )

Run each elements process method

fn process_mass( &mut self, _element_grid_conv_neigh: &mut ElementGridConvolutionNeighbors )

Process the heat of the grid Currently disabled as it is broken Process the mass of the grid and the mass above the grid

impl ElementGrid

pub fn get_texture(&self) -> RawImage

Draw the texture as the color of each element