Module orbiting_sand::physics::fallingsand::elements::element

This module contains the trait Element and associated types This is the trait that all elements must implement It also contains info about states of matter and other useful enums and components

Structs

Density

The density of the element relative to the cell width In units of kg/m^2

ElementTypeIter

An iterator over the variants of ElementType

SetHeatOnZeroSpecificHeatError

If something has 0 heat capacity or specific heat, you should not set its heat

StateOfMatterIter

An iterator over the variants of StateOfMatter

Enums

ElementTakeOptions

What to do after process is called on the elementgrid The element grid takes the element out of the grid so that it can’t self reference in the process operation for thread safety. However, we shouldn’t put it back if the element has moved, instead we will ask the element itself to clone itself and put the clone somewhere else

ElementType

Allows you to match on the type of element each element impl has a unique item in this enum

StateOfMatter

Traits

Element

This is the trait that all elements must implement