Trait orbiting_sand::physics::fallingsand::elements::element::Element

pub trait Element: Send + Sync {
    // Required methods
    fn get_type(&self) -> ElementType;
    fn get_last_processed(&self) -> Clock;
    fn get_color(&self) -> Color;
    fn get_density(&self) -> Density;
    fn get_state_of_matter(&self) -> StateOfMatter;
    fn box_clone(&self) -> Box<dyn Element>;
    fn _process(
        &mut self,
        pos: JkVector,
        coord_dir: &CoordinateDir,
        target_chunk: &mut ElementGrid,
        element_grid_conv: &mut ElementGridConvolutionNeighbors,
        current_time: Clock
    ) -> ElementTakeOptions;
    fn _set_last_processed(&mut self, current_time: Clock);

    // Provided methods
    fn get_mass(&self, cell_width: Length) -> Mass { ... }
    fn process(
        &mut self,
        pos: JkVector,
        coord_dir: &CoordinateDir,
        target_chunk: &mut ElementGrid,
        element_grid_conv: &mut ElementGridConvolutionNeighbors,
        current_time: Clock
    ) -> ElementTakeOptions { ... }
    fn try_swap_me(
        &self,
        pos1: ConvolutionIdx,
        target_chunk: &mut ElementGrid,
        element_grid_conv: &mut ElementGridConvolutionNeighbors,
        current_time: Clock
    ) -> ElementTakeOptions { ... }
}

This is the trait that all elements must implement

Required Methods

fn get_type(&self) -> ElementType

This gets the type of the element Converts between the trait and the enum

fn get_last_processed(&self) -> Clock

This gets the last time the element was processed Useful for physics calculations by getting the dt between now and then

fn get_color(&self) -> Color

This gets the color of the element Always constant and unique for each element, so that we can process them in fragment shaders knowing their type just by their color You can map them to other colors and add effects using the fragment shader

fn get_density(&self) -> Density

This gets the density of the element relative to the cell_width This is so bigger cells have more mass, so we don’t have to have as many cells for simpler bodies, like gas giants or the sun

fn get_state_of_matter(&self) -> StateOfMatter

This gets the state of matter of the element

fn box_clone(&self) -> Box<dyn Element>

This is the way we implement clone for a trait object

fn _process( &mut self, pos: JkVector, coord_dir: &CoordinateDir, target_chunk: &mut ElementGrid, element_grid_conv: &mut ElementGridConvolutionNeighbors, current_time: Clock ) -> ElementTakeOptions

This is the private process method to be implemented by the element Takes in the position of the element in the grid The coordinate grid which gives us information about the celestial body The target chunk which is the element grid this element was a part of The element grid convolution neighbors which gives you the ability to move and look around at neighboring chunks The current time

fn _set_last_processed(&mut self, current_time: Clock)

Set the last time the element was processed No need to call this publicly, it is called by the public process method

Provided Methods

fn get_mass(&self, cell_width: Length) -> Mass

This gets the mass of the element based on the density and the cell_width

fn process( &mut self, pos: JkVector, coord_dir: &CoordinateDir, target_chunk: &mut ElementGrid, element_grid_conv: &mut ElementGridConvolutionNeighbors, current_time: Clock ) -> ElementTakeOptions

This is the “public” process method, that calls the private _process method makes sure that _set_last_processed is called

fn try_swap_me( &self, pos1: ConvolutionIdx, target_chunk: &mut ElementGrid, element_grid_conv: &mut ElementGridConvolutionNeighbors, current_time: Clock ) -> ElementTakeOptions

Instructs the loop to swap the element with the element at pos1 you should have already checked to see if pos1 is valid, most likely it comes from another function as such this function will panic if pos1 is invalid