Struct orbiting_sand::physics::fallingsand::mesh::chunk_coords::ChunkCoords

pub struct ChunkCoords {
    width: Length,
    chunk_idx: ChunkIjkVector,
    start_concentric_circle_layer_relative: usize,
    start_concentric_circle_absolute: usize,
    start_radial_line: usize,
    end_radial_line: usize,
    layer_num_radial_lines: usize,
    num_concentric_circles: usize,
}

This is a chunk that represents a “full” layer. It doesn’t split itself in either the radial or concentric directions.

Fields
width: Lengthchunk_idx: ChunkIjkVectorstart_concentric_circle_layer_relative: usizestart_concentric_circle_absolute: usizestart_radial_line: usizeend_radial_line: usizelayer_num_radial_lines: usizenum_concentric_circles: usize
Implementations

impl ChunkCoords

pub fn get_positions(&self, mode: VertexMode) -> Vec<Vec2>

pub fn get_outline(&self) -> Vec<Vec2>

Similar to get_circle_vertexes, but the j index just iterates on the 0th and last element

pub fn get_bounding_box(&self) -> Rect

Gets the min and max positions in raw x, y of the chunk

pub fn get_uvs(&self, mode: VertexMode) -> Vec<Vec2>

Gets the UV coordinates of the vertexes of the chunk This is a more traditional square grid If you set skip to 1, you will get the full resolution If you set skip to 2, you will get half the resolution …

pub fn get_indices(&self, mode: VertexMode) -> Vec<u32>

impl ChunkCoords

pub fn total_size(&self) -> usize

pub fn get_cell_width(&self) -> Length

pub fn get_start_radius(&self) -> f32

pub fn get_end_radius(&self) -> f32

pub fn get_num_radial_lines(&self) -> usize

pub fn get_num_concentric_circles(&self) -> usize

pub fn get_end_radial_theta(&self) -> f32

pub fn get_start_radial_theta(&self) -> f32

pub fn get_start_concentric_circle_layer_relative(&self) -> usize

pub fn get_start_concentric_circle_absolute(&self) -> usize

pub fn get_end_concentric_circle_absolute(&self) -> usize

pub fn get_end_concentric_circle_layer_relative(&self) -> usize

pub fn get_end_radial_line(&self) -> usize

pub fn get_start_radial_line(&self) -> usize

pub fn get_layer_num(&self) -> usize

pub fn get_chunk_idx(&self) -> ChunkIjkVector

pub fn contains(&self, idx: IjkVector) -> bool

Checks to see if an absolute position around the circle is in the chunk

pub fn get_internal_coord_from_external_coord( &self, external_coord: IjkVector ) -> JkVector

Converts a coordinate from anywhere on the circle, assuming it is in the chunk to a coordinate inside the grid of this chunk

pub fn get_external_coord_from_internal_coord( &self, internal_coord: JkVector ) -> IjkVector

Converts a coordinate from inside this chunk to a coordinate on the circle

pub fn get_vertices(&self, mode: VertexMode) -> Vec<Vertex>

pub fn calc_chunk_outline(&self) -> OwnedMeshData

pub fn calc_chunk_meshdata(&self) -> OwnedMeshData

pub fn calc_chunk_triangle_wireframe(&self) -> OwnedMeshData

pub fn rel_pos_to_cell_idx( &self, xy_coord: RelXyPoint ) -> Result<IjkVector, String>

Converts a position relative to the origin of the circle to a cell index Returns an Err if the position is not on the circle

pub fn absolute_cell_idx_to_in_chunk_cell_idx( &self, cell_idx: IjkVector ) -> Result<JkVector, String>

Convert a cell coordinate “on the circle” to a position “on the chunk” Return an Err if this is not on the chunk