Module orbiting_sand::physics::fallingsand::util::functions

Functions

grid_iter

This is like the “skip” method but it always keeps the first and last item If it is larger than the number of items, it will just return the first and last item If the step is not a multiple of the number of items, it will round down to the previous multiple

interpolate_points

Finds a point halfway between two points

is_pow_2

Tests if a number is a power of 2 I found it’s important that some values are powers of two in order to enable grid_iter to work

modulo

A modulo that works for negative numbers

valid_step

Tests if a step is valid for a grid_iter A valid step is 1, len - 1, or a factor of len - 1 We convert things less than 1 to 1, or things greater than len - 1 to len - 1