Module orbiting_sand::physics::fallingsand::util::grid

A simple 2d grid type This was originally created seperate from the ndarray crate, but it was later decided to use the ndarray crate as the backend for this type. This is because the ndarray crate has a convolution function that is helpful for the physics simulation. So some of this code is now redundant, but is maintained for legacy reasons

Structs

Grid

A simple 2d grid type

GridOutOfBoundsError

Defines when the user has simply exceeded the bounds of the convolution

Functions

filter_vecgrid

Where filter is true, get the textures