Module orbiting_sand::physics::fallingsand::util::vectors

A collection of coordinate types and their conversions Mostly for the ChunkCoords crate::physics::fallingsand::mesh::coordinate_dir::CoordinateDir

Structs

ChunkIjkVector

The IjkVectorof a chunk within an crate::physics::fallingsand::data::element_directory::ElementGridDir In this case Ijk relate to the index of the chunk itself, not perportional to the cells within the chunk jk vector

FullIdx

Defines both the chunk and the internal idx of the element

IjkVector

Same as JkVector but with i indicating the “layer number” The core is layer 0 jk vector

JkVector

My personal coordinate type for the circular grids basically radius-theta coordinates, with integer radius and theta “counter clockwise” is positive just like in the unit circle

NdArrayCoords

A coordinate system for ndarray ndarray is row-major, so the jk vector is flipped Top left is (0, 0) ndarray coords

RelJkVector

This defines a movement or a vector relative to some position on the circular grid Same as JkVector, but with isize type fields which can contain negative numbers jk vector

TempJkVector

Sometimes while resolving a relative JkVector into a JkVector when you need isize type fields jk vector

Vertex

A vertex in a mesh Originally from ggez TODO: move to bevy’s types