Module orbiting_sand::physics::heat::components

Most of the units of measure as well as bevy components for heat related physics.

Structs

Area

The area of a system in square meters.

HeatCapacity

The amount of heat energy required to raise the temperature of a system. Measured in joules per degree celsius.

HeatEnergy

The amount of heat energy in a system. Measured in joules.

Length

The length of a system in meters.

SpecificHeat

The amount of heat energy required to raise 1 kilogram of the substance by one degree. Measured in joules per kilogram per degree celsius.

ThermalConductivity

Thermal Conductivity The ability of a material to conduct heat. Measured in watts per meter per degree celsius.

ThermodynamicTemperature

The temperature of a system. Measured in Kelvin. This should be represented in “real” units, physically similar to the real world.

Constants

ROOM_TEMPERATURE_K

The temperature in kelvin of room temperature.