Module orbiting_sand::physics::orbits::nbody

The math functions for orbits and orbital mechanics.

NxM-Body Physics

In this module, we implement a type of gravitational physics simulation I will call “NxM-Body” physics.

Similar to N-Body physics, this models a more complex system of gravity than mere two-body physics. However, it is not as complex as N-Body physics, which models the gravitational interactions between all bodies in the system. Instead, NxM-Body physics models the gravitational interactions between all bodies in the system that have a gravitational field, which will likely just be the planets, moons, and stars, likely less than 100 bodies. We can call this N. The other bodies, which will likely be spacecraft, asteroids, and other small bodies, will be modeled as if they are affected by the gravitational fields of the bodies with gravitational fields, but not by each other. We can call this M.

This is a simple model that is much less computationally expensive than N-Body physics, which is worst-case O(N^2), wheras NxM-Body physics is worst-case O(N*M). Since M is constant and small in any given simulation, this is effectively O(N).

Leapfrog Integration

https://en.wikipedia.org/wiki/Leapfrog_integration

We use the leapfrog integration method to update the positions and velocities of the bodies. This is a symplectic integrator , which means it conserves energy. This is important for a physics simulation, because it means that the simulation will not “drift” over time, and the bodies will not gain or lose energy over time.

It is also a second-order integrator, which means that it is more accurate than a first-order integrator, like Euler’s method. This means that the simulation will be more accurate, and will not require as small of a time step to be stable.

It is acomplished by updating the velocity of the bodies by half a time step, then updating the position of the bodies by a full time step, then updating the velocity of the bodies by half a time step again.

Decision to use the CPU

I originally wanted to use the GPU for this simulation, because it is a very parallel problem. However, I decided to use the CPU because it is a simpler solution, and each frame I would need to read the positions and velocities of all the bodies from the GPU which is actually slower than just doing the computation on the CPU. Using rayon, I can parallelize the computation on the CPU, so it is not a huge performance hit.

Structs

NBodyPlugin

Plugin to set up nbody physics

Constants

G

The gravitational constant

MIN_DISTANCE_SQUARED

It’s important that we don’t compute the gravitational force between two bodies that are too close together, because the force will be very large and the simulation will be unstable. This is squared because then we don’t have to do a square root to compare it to the distance squared in the force calculation.

Functions

compute_gravitational_force 🔒

Returns the gravitational force between two entities

full_position_update 🔒

Updates the position of the entity one full step

half_step_velocity_update 🔒

Updates the velocity of the entity one half step