Module orbiting_sand::physics::fallingsand::convolution::neighbor_indexes

Indexes in ChunkIjkVectors for all the neighbors of a chunk

Structs

BottomNeighborIdxsIter

An iterator for the bottom neighbor indexes

ElementGridConvolutionNeighborIdxs

The main type exported by this module Contains all the ChunkIjkVector indexes for the convolution Check out the super::neighbor_identifiers::ConvolutionIdentifier and super::neighbor_indexes::ElementGridConvolutionNeighborIdxs documentation for more information

ElementGridConvolutionNeighborIdxsIter

An iterator for the neighbor indexes

LeftRightNeighborIdxsIter

An iterator for the left and right neighbor indexes

TopNeighborIdxsIter

An iterator for the top neighbor indexes

Enums

BottomNeighborIdxs

Bottom neighbor indexes in the convolution Check out the super::neighbor_identifiers::BottomNeighborIdentifier and super::neighbor_grids::BottomNeighborGrids documentation for more information

LeftRightNeighborIdxs

Left and Right neighbor indexes in the convolution Check out the super::neighbor_identifiers::LeftRightNeighborIdentifier and super::neighbor_grids::LeftRightNeighborGrids documentation for more information documentation for more information

TopNeighborIdxs

Top neighbor indexes in the convolution Check out the super::neighbor_identifiers::TopNeighborIdentifier and super::neighbor_grids::TopNeighborGrids documentation for more information