Module orbiting_sand::physics::fallingsand::util::vectors

A collection of coordinate types and their conversions Mostly for the ChunkCoords crate::physics::fallingsand::mesh::coordinate_directory::CoordinateDir

Structs

ChunkIjkVector

The IjkVector of a chunk within an crate::physics::fallingsand::data::element_directory::ElementGridDir In this case Ijk relate to the index of the chunk itself, not perportional to the cells within the chunk

FullIdx

Defines both the chunk and the internal idx of the element

IjkVector

Same as JkVector, but with i indicating the “layer number” The core is layer 0

JkVector

My personal coordinate type for the circular grids basically radius-theta coordinates, with integer radius and theta “counter clockwise” is positive just like in the unit circle j is the “concentric circle” or “radial” axis, kinda like y, towards the core is 0 k is the “tangential” axis, kinda like x, positive is counter clockwise from unit circle 0 degrees which is starting from 3 o’clock east Can also be used to describe a grid, like a chunk taken from the circle In this case j is the height and k is the width Bottom right is (0, 0) If you need to also know the layer number, use IjkVector If you need a relative vector, use RelJkVector

NdArrayCoords

A coordinate system for ndarray ndarray is row-major, so the jk vector is flipped Top left is (0, 0)

RelJkVector

This defines a movement or a vector relative to some position on the circular grid Same as JkVector, but with isize type fields which can contain negative numbers

TempJkVector

Sometimes while resolving a relative JkVector into a JkVector when you need isize type fields

Vertex

A vertex in a mesh Originally from ggez TODO: move to bevy’s types